rename util.ts -> index.ts
